Made design changes
Looks better on a phone, sections labelled as the meeting notes said,
CSS updated but still more inline to add, country selection added but
not functional, map still needs to have functionality removed, social
media icons added but not linked and color might not be final, footer
details added and links linked, search toggle needs adjustment as it no
longer is flush with map still waiting on options for that. NOTE that
none of the new features or old features have been hooked up to the API
yet map still needs work.